Change Transction Receipt to interface
