+++ yggdrash-core/src/test/java/io/yggdrash/core/contract/TransactionReceiptTest.java
@@ -27,7 +27,7 @@ public class TransactionReceiptTest {
 
     @Before
     public void setUp() {
-        txReceipt = new TransactionReceipt();
+        txReceipt = new TransactionReceiptImpl();
     }
 
     @Test
@@ -39,7 +39,7 @@ public class TransactionReceiptTest {
     @Test
     public void statusTest() {
         assert !txReceipt.isSuccess();
-        txReceipt.setStatus(TransactionReceipt.SUCCESS);
+        txReceipt.setStatus(ExecuteStatus.SUCCESS);
         assert txReceipt.isSuccess();
     }
